Migrate entity types to util library
