+++ unified-codes/libs/util/src/lib/types/EntityNode.ts
@@ -1,0 +1,8 @@
+import { Property } from "./Property";
+
+export type EntityNode = {
+    code: string;
+    description: string;
+    type: string;
+    properties?: Property[];
+};
