+++ unified-codes/libs/util/src/lib/types/Property.ts
